% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{address_unpack}
\alias{address_unpack}
\title{Unpack repo address}
\usage{
address_unpack(repo)
}
\arguments{
\item{repo}{Repository address}
}
\value{
list
}
\description{
Returns a list of all the elements that make up a repository
address (service, username, repo, ref).
}
