% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{all_search}
\alias{all_search}
\title{Search for outsider modules}
\usage{
all_search()
}
\value{
data.frame
}
\description{
Returns GitHub API item results for outsider module search.
}
\seealso{
Other private-search: \code{\link{build_status}},
  \code{\link{read_yaml}}, \code{\link{repo_search}},
  \code{\link{tags}}, \code{\link{yaml}}
}
\concept{private-search}
