% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{available}
\alias{available}
\title{Look up information on all available outsider modules}
\usage{
available()
}
\value{
data.frame
}
\description{
Return information on all available outsider modules
}
\seealso{
Other user: \code{\link{log_set}},
  \code{\link{module_exists}}, \code{\link{module_help}},
  \code{\link{module_import}},
  \code{\link{module_installed}},
  \code{\link{module_install}}, \code{\link{module_test}},
  \code{\link{module_uninstall}}
}
\concept{user}
