% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitbucket.R
\name{bitbucket_repo_search}
\alias{bitbucket_repo_search}
\title{Search for repository}
\usage{
bitbucket_repo_search(repo)
}
\arguments{
\item{repo}{bitbucket repo}
}
\value{
data.frame
}
\description{
Return bitbucket API item for specific repository.
}
