% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{build_status}
\alias{build_status}
\title{Look-up details on program}
\usage{
build_status(repo)
}
\arguments{
\item{repo}{GitHub repo}
}
\value{
Logical
}
\description{
Is build passing? Returns either TRUE or FALSE.
}
\seealso{
Other private: \code{\link{default_log_set}},
  \code{\link{fnames_get}},
  \code{\link{is_docker_available}},
  \code{\link{is_filepath}}, \code{\link{log_get}},
  \code{\link{om_search}}, \code{\link{om_yaml}},
  \code{\link{test_examples}}, \code{\link{test_import}},
  \code{\link{test_install}}
}
\concept{private}
