% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{copy.container}
\alias{copy.container}
\title{Copy files to and from a docker container}
\usage{
\method{copy}{container}(x, send = NULL, rtrn = NULL)
}
\arguments{
\item{x}{Container ID}

\item{send}{Filepaths to send from host computer to container.}

\item{rtrn}{Directory on host computer where returning files should be sent.}
}
\value{
Logical
}
\description{
Copy all given host files to a running docker container or
copy all files from a container's working directory to a host filepath.
}
\details{
All outsider modules have a working_dir/ in which generated files
are created and initiation files must be for the program to use.
Files must be sent to this working directory and then returned before and
after the program has run.

If no \code{send} or \code{rtrn} specified, returns TRUE.
}
