% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{default_log_set}
\alias{default_log_set}
\title{Set default log streams}
\usage{
default_log_set()
}
\description{
By default all streams are printed to console with the exception
of \code{docker_out}.
}
\seealso{
Other private: \code{\link{build_status}},
  \code{\link{examples_test}}, \code{\link{fnames_get}},
  \code{\link{import_test}}, \code{\link{install_test}},
  \code{\link{is_docker_available}},
  \code{\link{is_filepath}}, \code{\link{log_get}},
  \code{\link{module_versions}}, \code{\link{module_yaml}}
}
\concept{private}
