% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_build}
\alias{docker_build}
\title{Build a docker image}
\usage{
docker_build(img_id, url)
}
\arguments{
\item{img_id}{Image ID}

\item{url}{Dockerfile URL}
}
\value{
Logical
}
\description{
Runs run \code{build} command.
}
