% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_cmd}
\alias{docker_cmd}
\title{Run docker command}
\usage{
docker_cmd(args, std_out = TRUE, std_err = TRUE)
}
\arguments{
\item{args}{Vector of arguments}

\item{std_out}{if and where to direct child process STDOUT.
See \code{\link[sys]{exec}}.}

\item{std_err}{if and where to direct child process STDERR.
See \code{\link[sys]{exec}}.}
}
\value{
Logical
}
\description{
Runs a docker command with provided arguments
}
\seealso{
Other private-docker: \code{\link{container-class}},
  \code{\link{docker_build}}, \code{\link{docker_cp}},
  \code{\link{docker_img_rm}},
  \code{\link{docker_ps_count}}
}
\concept{private-docker}
