% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_cp}
\alias{docker_cp}
\title{Copy files to and from container}
\usage{
docker_cp(origin, dest)
}
\arguments{
\item{origin}{Origin filepath}

\item{dest}{Destination filepath}
}
\value{
Logical
}
\description{
Copy files to and from running Docker container
}
\details{
Container foldes are indicated with
\code{[container_id]:[filepath]}.
}
