% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_img_rm}
\alias{docker_img_rm}
\title{Remove docker image}
\usage{
docker_img_rm(img)
}
\arguments{
\item{img}{Image name}
}
\value{
Logical
}
\description{
Deletes docker image from system.
}
\seealso{
Other private-docker: \code{\link{container-class}},
  \code{\link{docker_build}}, \code{\link{docker_cmd}},
  \code{\link{docker_cp}}, \code{\link{docker_img_ls}},
  \code{\link{docker_ps_count}}, \code{\link{docker_pull}}
}
\concept{private-docker}
