% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_killall}
\alias{docker_killall}
\title{Attempt to kill all running docker containers}
\usage{
docker_killall()
}
\value{
Logical
}
\description{
In the event a user loses track of the number of docker 
containers they have created, this function will stop and remove all
active and inactive containers.
}
\details{
If you are running any non-outsider contianers, use this function
with caution.
}
\seealso{
Other private-docker: \code{\link{container-class}},
  \code{\link{docker_build}}, \code{\link{docker_cmd}},
  \code{\link{docker_cp}}, \code{\link{docker_img_rm}},
  \code{\link{docker_ps_count}}
}
\concept{private-docker}
