% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_pull}
\alias{docker_pull}
\title{Pull an image from DockerHub.}
\usage{
docker_pull(img, tag = "latest")
}
\arguments{
\item{img}{Image name}

\item{tag}{Tag version, default latest.}
}
\value{
Logical
}
\description{
Speeds up outsider module installation by downloading compiled
images.
}
\seealso{
Other private-docker: \code{\link{container-class}},
  \code{\link{docker_build}}, \code{\link{docker_cmd}},
  \code{\link{docker_cp}}, \code{\link{docker_img_ls}},
  \code{\link{docker_img_rm}},
  \code{\link{docker_ps_count}}
}
\concept{private-docker}
