% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{.arglist_get}
\alias{.arglist_get}
\title{Generate vector of arguments}
\usage{
.arglist_get(...)
}
\arguments{
\item{...}{Any number of arguments}
}
\value{
Character vector
}
\description{
Convert all the arguments passed to this function, including
those contained in '...', into character vector.
}
