% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{.arglist_parse}
\alias{.arglist_parse}
\title{Normalise arguments for docker container}
\usage{
.arglist_parse(arglist, keyvals_to_drop = NULL, vals_to_drop = NULL,
  normalise_paths = TRUE)
}
\arguments{
\item{arglist}{Arguments as character vector}

\item{keyvals_to_drop}{Argument keys to drop, e.g. -wd.}

\item{vals_to_drop}{Specific values to drop, e.g. --verbose.}
}
\value{
Character vector
}
\description{
Utility function for parsing the arguments provided by a user.
Drop any specified key:value pairs with \code{keyvals_to_drop} or drop any
specific values \code{vals_to_drop}. With \code{normalise_paths} as TRUE,
all filepaths in the arglist will be converted to basenames.
}
\seealso{
Other developer: \code{\link{.arglist_get}},
  \code{\link{.filestosend_get}},
  \code{\link{.travis_yaml}}, \code{\link{.wd_get}},
  \code{\link{module_check}},
  \code{\link{module_identities}},
  \code{\link{module_skeleton}}, \code{\link{module_test}},
  \code{\link{outsider-class}}
}
\concept{developer}
