% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{.args_parse}
\alias{.args_parse}
\title{Parse arguments}
\usage{
.args_parse(n = 1L)
}
\arguments{
\item{n}{Number of generations to go back, integer}
}
\value{
Character vector
}
\description{
Convert '...' of the function that calls this function into
an evaluated, character vector.
}
\details{
Parse arguments by converting all the arguments provided by the
function that calls this function. E.g. foo(...) calls .args_parse(). All
arguments provided in foo() as '...' are evaluated and converted into
a character vector.
}
\seealso{
Other developer: \code{\link{.copy_from_docker}},
  \code{\link{.copy_to_docker}},
  \code{\link{.docker_build}}, \code{\link{.docker_cmd}},
  \code{\link{.docker_cp}}, \code{\link{.docker_exec}},
  \code{\link{.docker_img_rm}},
  \code{\link{.docker_ps_count}},
  \code{\link{.docker_start}}, \code{\link{.docker_stop}},
  \code{\link{.ids_get}}, \code{\link{.pkgnm_to_prgm}},
  \code{\link{.pkgnm_to_repo}}, \code{\link{.repo_to_img}},
  \code{\link{.repo_to_pkgnm}}, \code{\link{.run}},
  \code{\link{.to_basename}}, \code{\link{.travisyml_gen}},
  \code{\link{.which_args_are_filepaths}}
}
\concept{developer}
