% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{.copy_from_docker}
\alias{.copy_from_docker}
\title{Copy all contents from working dir}
\usage{
.copy_from_docker(cntnr_id)
}
\arguments{
\item{cntnr_id}{Container ID}
}
\value{
Logical
}
\description{
Copy all the contents of the working_dir on the outsider Docker
container to the host machine.
}
\details{
All outsider modules have a working_dir/ when generated files are
created. These must be copied from the container to the host machine
for the user to interact with.
}
\seealso{
Other developer: \code{\link{.args_parse}},
  \code{\link{.copy_to_docker}},
  \code{\link{.docker_build}}, \code{\link{.docker_cmd}},
  \code{\link{.docker_cp}}, \code{\link{.docker_exec}},
  \code{\link{.docker_img_rm}},
  \code{\link{.docker_ps_count}},
  \code{\link{.docker_start}}, \code{\link{.docker_stop}},
  \code{\link{.ids_get}}, \code{\link{.pkgnm_to_prgm}},
  \code{\link{.pkgnm_to_repo}}, \code{\link{.repo_to_img}},
  \code{\link{.repo_to_pkgnm}}, \code{\link{.run}},
  \code{\link{.travisyml_gen}},
  \code{\link{.which_args_are_filepaths}}
}
\concept{developer}
