% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{.copy_to_docker}
\alias{.copy_to_docker}
\title{Copy files to docker container}
\usage{
.copy_to_docker(cntnr_id, host_flpths)
}
\arguments{
\item{cntnr_id}{Container ID}

\item{host_flpths}{Filepaths to send on host computer}
}
\value{
Logical
}
\description{
Copy all given host files to a running docker container.
}
\seealso{
Other developer: \code{\link{.args_parse}},
  \code{\link{.copy_from_docker}},
  \code{\link{.docker_build}}, \code{\link{.docker_cmd}},
  \code{\link{.docker_cp}}, \code{\link{.docker_exec}},
  \code{\link{.docker_img_rm}},
  \code{\link{.docker_ps_count}},
  \code{\link{.docker_start}}, \code{\link{.docker_stop}},
  \code{\link{.ids_get}}, \code{\link{.pkgnm_to_prgm}},
  \code{\link{.pkgnm_to_repo}}, \code{\link{.repo_to_img}},
  \code{\link{.repo_to_pkgnm}}, \code{\link{.run}},
  \code{\link{.to_basename}}, \code{\link{.travisyml_gen}},
  \code{\link{.which_args_are_filepaths}}
}
\concept{developer}
