% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{.docker_build}
\alias{.docker_build}
\title{Build a docker image}
\usage{
.docker_build(img_id, url)
}
\arguments{
\item{img_id}{Image ID}

\item{url}{Dockerfile URL}
}
\value{
Logical
}
\description{
Runs run \code{build} command.
}
\seealso{
Other developer: \code{\link{.args_parse}},
  \code{\link{.docker_cmd}}, \code{\link{.docker_cp}},
  \code{\link{.docker_exec}}, \code{\link{.docker_img_rm}},
  \code{\link{.docker_ps_count}},
  \code{\link{.docker_start}}, \code{\link{.docker_stop}},
  \code{\link{.ids_get}}, \code{\link{.pkgnm_to_prgm}},
  \code{\link{.pkgnm_to_repo}}, \code{\link{.repo_to_img}},
  \code{\link{.repo_to_pkgnm}}, \code{\link{.run}},
  \code{\link{.which_args_are_filepaths}}
}
\concept{developer}
