% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{.docker_cmd}
\alias{.docker_cmd}
\title{Run docker command}
\usage{
.docker_cmd(args, std_out = TRUE, std_err = TRUE)
}
\arguments{
\item{args}{Vector of arguments}

\item{std_out}{if and where to direct child process STDOUT.
See \code{\link[sys]{exec}}.}

\item{std_err}{if and where to direct child process STDERR.
See \code{\link[sys]{exec}}.}
}
\value{
Logical
}
\description{
Runs a docker command with provided arguments
}
\seealso{
Other developer: \code{\link{.args_parse}},
  \code{\link{.docker_build}}, \code{\link{.docker_cp}},
  \code{\link{.docker_exec}}, \code{\link{.docker_img_rm}},
  \code{\link{.docker_ps_count}},
  \code{\link{.docker_start}}, \code{\link{.docker_stop}},
  \code{\link{.ids_get}}, \code{\link{.pkgnm_to_prgm}},
  \code{\link{.pkgnm_to_repo}}, \code{\link{.repo_to_img}},
  \code{\link{.repo_to_pkgnm}}, \code{\link{.run}},
  \code{\link{.which_args_are_filepaths}}
}
\concept{developer}
