% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{.docker_exec}
\alias{.docker_exec}
\title{Execute in a docker container}
\usage{
.docker_exec(cntnr_id, ...)
}
\arguments{
\item{cntnr_id}{Container ID}

\item{...}{Command and arguments for container.}
}
\value{
Logical
}
\description{
Pass commands to a running docker container with the \code{exec}
command.
}
