% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{.docker_start}
\alias{.docker_start}
\title{Start a docker container}
\usage{
.docker_start(cntnr_id, img_id)
}
\arguments{
\item{cntnr_id}{Container ID}

\item{img_id}{Image ID}
}
\value{
Logical
}
\description{
Runs run \code{run} command on a docker image.
}
