% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{.docker_stop}
\alias{.docker_stop}
\title{Stop and delete a docker container}
\usage{
.docker_stop(cntnr_id)
}
\arguments{
\item{cntnr_id}{Container ID}
}
\value{
Logical
}
\description{
Runs run \code{stop} and \code{rm} commands on running docker
containers.
}
