% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{.ids_get}
\alias{.ids_get}
\title{Get image ID and unique container ID}
\usage{
.ids_get(pkgnm)
}
\arguments{
\item{pkgnm}{Package name}
}
\value{
Logical
}
\description{
From a pkgnm, return the image ID and a unique container ID.
}
