% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{.module_identities}
\alias{.module_identities}
\title{Return identities for a module}
\usage{
.module_identities(flpth)
}
\arguments{
\item{flpth}{File path to location of module}
}
\value{
Logical
}
\description{
Returns a list of the identities (GitHub repo, Package name,
Docker images) for an outsider module. Works for modules in development.
Requires module to have a file path.
}
\seealso{
Other developer: \code{\link{.arglist_get}},
  \code{\link{.arglist_parse}}, \code{\link{.dirpath_get}},
  \code{\link{.filestosend_get}},
  \code{\link{.module_check}},
  \code{\link{.module_skeleton}},
  \code{\link{.module_test}}, \code{\link{.module_travis}},
  \code{\link{.wd_get}}, \code{\link{outsider-class}}
}
\concept{developer}
