% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{.module_skeleton}
\alias{.module_skeleton}
\title{Generate a skeleton for a module}
\usage{
.module_skeleton(program_name, github_user, docker_user, flpth = getwd())
}
\arguments{
\item{program_name}{Name of the command-line program}

\item{github_user}{Developer's username for GitHub}

\item{docker_user}{Developer's username for Docker}

\item{flpth}{File path to location of where module will be created, default
current working directory.}
}
\value{
Logical
}
\description{
Create all the base files and folders to kickstart the
development of a new outsider module.
}
\details{
Module developers must have a GitHub and Docker-Hub account.
For more detailed information, see online.
}
\seealso{
Other developer: \code{\link{.arglist_get}},
  \code{\link{.arglist_parse}},
  \code{\link{.filestosend_get}},
  \code{\link{.module_check}},
  \code{\link{.module_identities}},
  \code{\link{.module_travis}}, \code{\link{.wd_get}},
  \code{\link{module_test}}, \code{\link{outsider-class}}
}
\concept{developer}
