% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{.module_test}
\alias{.module_test}
\title{Test an outsider module}
\usage{
.module_test(repo, verbose = FALSE)
}
\arguments{
\item{repo}{Module repo}

\item{verbose}{Print docker and program info to console}
}
\value{
Logical
}
\description{
Ensure an outsider module builds, imports correctly and all
its functions successfully complete.
}
\details{
Success or fail, the module is uninstalled from the machine after
the test is run.
}
\examples{
library(outsider)
repo <- 'dombennett/om..hello.world'
# installs, checks functions, runs function examples, uninstalls
# module must already be uploaded to GitHub and Docker Hub
# NOT RUN
# .module_test(repo = repo, verbose = TRUE)
}
\seealso{
Other developer: \code{\link{.arglist_get}},
  \code{\link{.arglist_parse}}, \code{\link{.dirpath_get}},
  \code{\link{.filestosend_get}},
  \code{\link{.module_check}},
  \code{\link{.module_identities}},
  \code{\link{.module_skeleton}},
  \code{\link{.module_travis}}, \code{\link{.wd_get}},
  \code{\link{outsider-class}}
}
\concept{developer}
