% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outsider.R
\name{.outsider_class}
\alias{.outsider_class}
\alias{.outsider_init}
\title{Construct outsider object}
\usage{
.outsider_init(repo, cmd = NA, arglist = NULL, wd = NULL,
  files_to_send = NULL)
}
\arguments{
\item{repo}{Repository of the outsider module}

\item{cmd}{Command to be called in the container}

\item{arglist}{Arguments for command, character vector}

\item{wd}{Directory to which program generated files will be returned}

\item{files_to_send}{Files to be sent to container}
}
\value{
A list of class \code{outsider} with the following items:
\item{repo}{Repository of the outsider module}
\item{pkgnm}{Package name of the outsider module}
\item{cmd}{Command to be called in the container}
\item{arglist}{Arguments for command, character vector}
\item{wd}{Directory to which program generated files will be returned}
\item{files_to_send}{Files to be sent to container}
\item{container}{Docker container object}
}
\description{
Returns a outsider object. The outsider object describes a
outsider module's program and arguments. The object is generated every
time an outsider module program is called. It details the arguments of a
call, the command as well as the files to send to the docker container.
}
\details{
The outsider module runs a docker container that acts like a 
separate machine on the host computer. All the files necessary for the 
program to be run must be sent to the remote machine before the program
is called.
The arguments, wd and files_to_send can all be defined after the outsider
has been initiated using \code{otsdr$``} notation.
Once a outsider has been defined, the command can be run using
\code{.run(otsdr)}.
The \code{arglist}, \code{wd} or \code{files_to_send} do not need to be
defined for the outsider to be run.
}
