% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{.pkgnm_to_prgm}
\alias{.pkgnm_to_prgm}
\title{Convert pkgnm to prgm}
\usage{
.pkgnm_to_prgm(pkgnm)
}
\arguments{
\item{pkgnm}{Package name.}
}
\value{
Logical
}
\description{
Converts the R package name of a module into its base program
name.
}
