% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{.pkgnm_to_repo}
\alias{.pkgnm_to_repo}
\title{Convert pkgnm to repo}
\usage{
.pkgnm_to_repo(pkgnm)
}
\arguments{
\item{pkgnm}{Package name}
}
\value{
Logical
}
\description{
Converts the R package name of a module into its repo name.
}
\seealso{
Other developer: \code{\link{.args_parse}},
  \code{\link{.docker_build}}, \code{\link{.docker_cmd}},
  \code{\link{.docker_cp}}, \code{\link{.docker_exec}},
  \code{\link{.docker_img_rm}},
  \code{\link{.docker_ps_count}},
  \code{\link{.docker_start}}, \code{\link{.docker_stop}},
  \code{\link{.ids_get}}, \code{\link{.pkgnm_to_prgm}},
  \code{\link{.repo_to_img}}, \code{\link{.repo_to_pkgnm}},
  \code{\link{.run}},
  \code{\link{.which_args_are_filepaths}}
}
\concept{developer}
