% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{.run}
\alias{.run}
\title{Run a command from a docker image}
\usage{
.run(pkgnm, files_to_send, ...)
}
\arguments{
\item{pkgnm}{Package name}

\item{files_to_send}{Filepaths on host to send to module container}

\item{...}{Command and arguments}
}
\value{
Logical
}
\description{
Run a command with arguments within a docker image. The function
will start a container, run the command and after completion stop and remove
the container.
}
