% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{.to_basename}
\alias{.to_basename}
\title{Reduce to filepaths to basename}
\usage{
.to_basename(args)
}
\arguments{
\item{x}{Character vector}
}
\value{
Character vector
}
\description{
Return return a vector where all valid filepaths are converted
to file basenames. E.g. "dir1/dir2/text.file" is converted to "text.file"
}
\seealso{
Other developer: \code{\link{.args_parse}},
  \code{\link{.copy_from_docker}},
  \code{\link{.copy_to_docker}},
  \code{\link{.docker_build}}, \code{\link{.docker_cmd}},
  \code{\link{.docker_cp}}, \code{\link{.docker_exec}},
  \code{\link{.docker_img_rm}},
  \code{\link{.docker_ps_count}},
  \code{\link{.docker_start}}, \code{\link{.docker_stop}},
  \code{\link{.ids_get}}, \code{\link{.pkgnm_to_prgm}},
  \code{\link{.pkgnm_to_repo}}, \code{\link{.repo_to_img}},
  \code{\link{.repo_to_pkgnm}}, \code{\link{.run}},
  \code{\link{.travisyml_gen}},
  \code{\link{.which_args_are_filepaths}}
}
\concept{developer}
