% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{.travis_yaml}
\alias{.travis_yaml}
\title{Generate travis file}
\usage{
.travis_yaml(repo, dir = getwd())
}
\arguments{
\item{repo}{Repository}

\item{dir}{Directory in which to create .travis.yml}
}
\value{
Logical
}
\description{
Write .travis.yml to working directory.
}
\details{
All validated outsider modules must have a .travis.yml in their
repository. These .travis.yml must be generated using this function.
}
\seealso{
Other developer: \code{\link{.arglist_get}},
  \code{\link{.arglist_parse}},
  \code{\link{.filestosend_get}}, \code{\link{.wd_get}},
  \code{\link{outsider-class}}
}
\concept{developer}
