% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{.travisyml_gen}
\alias{.travisyml_gen}
\title{Generate travis file}
\usage{
.travisyml_gen(repo)
}
\arguments{
\item{repo}{Repository}
}
\value{
Logical
}
\description{
Write .travis.yml to working directory.
}
\details{
All validated outsider modules must have a .travis.yml in their
repository. These .travis.yml must be generated using this function.
}
\seealso{
Other developer: \code{\link{.args_parse}},
  \code{\link{.copy_from_docker}},
  \code{\link{.copy_to_docker}},
  \code{\link{.docker_build}}, \code{\link{.docker_cmd}},
  \code{\link{.docker_cp}}, \code{\link{.docker_exec}},
  \code{\link{.docker_img_rm}},
  \code{\link{.docker_ps_count}},
  \code{\link{.docker_start}}, \code{\link{.docker_stop}},
  \code{\link{.ids_get}}, \code{\link{.pkgnm_to_prgm}},
  \code{\link{.pkgnm_to_repo}}, \code{\link{.repo_to_img}},
  \code{\link{.repo_to_pkgnm}}, \code{\link{.run}},
  \code{\link{.to_basename}},
  \code{\link{.which_args_are_filepaths}}
}
\concept{developer}
