% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{.which_args_are_filepaths}
\alias{.which_args_are_filepaths}
\title{Determine which args are filepaths}
\usage{
.which_args_are_filepaths(args)
}
\arguments{
\item{args}{Character vector of arguments}
}
\value{
Character vector
}
\description{
Return filepaths from args.
}
