% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_and_install}
\alias{download_and_install}
\title{Download and install module}
\usage{
download_and_install(url, tag = "latest", pull = TRUE,
  verbose = TRUE)
}
\arguments{
\item{url}{.tar.gz link}

\item{tag}{Docker tag, default 'latest'}

\item{pull}{Attempt to pull from Docker Hub?}

\item{verbose}{Be verbose? Default TRUE.}
}
\value{
logical(1)
}
\description{
Download and build an outsider module. The function will
download the module to a temporary directory and then install.
}
\details{
\code{url} is expected to be a .tar.gz link.
}
