% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{file_create}
\alias{file_create}
\title{Create file}
\usage{
file_create(x, flpth)
}
\description{
Write x to a filepath. Forces creation of directories.
}
\seealso{
Other private: \code{\link{default_log_set}},
  \code{\link{examples_test}}, \code{\link{fnames_get}},
  \code{\link{import_test}}, \code{\link{install_test}},
  \code{\link{install}}, \code{\link{is_filepath}},
  \code{\link{is_installed}}, \code{\link{log_get}},
  \code{\link{string_replace}},
  \code{\link{templates_get}}, \code{\link{to_basename}}
}
\concept{private}
