% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{fnames_get}
\alias{fnames_get}
\title{Function names for module}
\usage{
fnames_get(repo)
}
\arguments{
\item{repo}{Module repo}
}
\value{
character vector
}
\description{
Return function names of all available functions for an
installed outsider modules
}
\seealso{
Other private: \code{\link{build_status}},
  \code{\link{default_log_set}},
  \code{\link{is_docker_available}}, \code{\link{log_get}},
  \code{\link{om_search}}, \code{\link{om_yaml}},
  \code{\link{test_examples}}, \code{\link{test_import}},
  \code{\link{test_install}}
}
\concept{private}
