% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_module_details}
\alias{github_module_details}
\title{Look up details of module(s) on GitHub}
\usage{
github_module_details(repo = NULL)
}
\arguments{
\item{repo}{Vector of one or more outsider module repositories, default NULL.}
}
\value{
tbl_df
}
\description{
Return a tbl_df of information for outsider module(s).
If \code{repo} is NULL, will return details on all available modules.
}
