% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_repo_search}
\alias{github_repo_search}
\title{Search for repository}
\usage{
github_repo_search(repo)
}
\arguments{
\item{repo}{GitHub repo}
}
\value{
data.frame
}
\description{
Return GitHub API item for specific repository.
}
