% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_tags}
\alias{github_tags}
\title{Module tags from GitHub}
\usage{
github_tags(repos)
}
\arguments{
\item{repos}{Character vector of outsider module repositories.}
}
\value{
tbl_df
}
\description{
Return tbl_df of module tags for a list of outsider
modules hosted on GitHub.
}
