% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_yaml}
\alias{github_yaml}
\title{Module YAML information}
\usage{
github_yaml(repos)
}
\arguments{
\item{repos}{Character vector of outsider module repositories on GitHub.}
}
\value{
tbl_df
}
\description{
Return tbl_df of all YAML information of given outsider
module repos.
}
