% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{gitlab_repo_search}
\alias{gitlab_repo_search}
\title{Search for repository}
\usage{
gitlab_repo_search(repo)
}
\arguments{
\item{repo}{gitlab repo}
}
\value{
data.frame
}
\description{
Return gitlab API item for specific repository.
}
