% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{gitlab_tags}
\alias{gitlab_tags}
\title{Module tags from GitLab}
\usage{
gitlab_tags(repo_ids)
}
\arguments{
\item{repo_ids}{Character vector of outsider module repositories ids.}
}
\value{
tbl_df
}
\description{
Return tbl_df of module tags for a list of outsider
modules hosted on gitlab.
}
