% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{ids_get}
\alias{ids_get}
\title{Get image ID and unique container ID}
\usage{
ids_get(pkgnm)
}
\arguments{
\item{pkgnm}{Package name}
}
\value{
Logical
}
\description{
From a pkgnm, return the image ID and a unique container ID.
}
\seealso{
Other private-ids: \code{\link{pkgnm_to_prgm}},
  \code{\link{pkgnm_to_repo}}, \code{\link{repo_to_img}},
  \code{\link{repo_to_pkgnm}}
}
\concept{private-ids}
