% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{import_test}
\alias{import_test}
\title{Test whether module functions can be imported}
\usage{
import_test(repo)
}
\arguments{
\item{repo}{Module repo}
}
\value{
logical
}
\description{
Return TRUE if all of the outsider module functions are
successfully imported.
}
\seealso{
Other private: \code{\link{build_status}},
  \code{\link{default_log_set}},
  \code{\link{examples_test}}, \code{\link{fnames_get}},
  \code{\link{install_test}},
  \code{\link{is_docker_available}},
  \code{\link{is_filepath}}, \code{\link{log_get}},
  \code{\link{module_versions}}, \code{\link{module_yaml}}
}
\concept{private}
