% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install module}
\usage{
install(repo, tag, dockerfile_url = NULL)
}
\arguments{
\item{repo}{GitHub repository name}

\item{tag}{Docker-Hub tag, e.g. 'latest'}

\item{dockerfile_url}{URL to Dockerfile}
}
\value{
logical
}
\description{
Build/pull Docker image and install R package
}
\seealso{
Other private: \code{\link{default_log_set}},
  \code{\link{examples_test}}, \code{\link{file_create}},
  \code{\link{fnames_get}}, \code{\link{import_test}},
  \code{\link{install_test}}, \code{\link{is_filepath}},
  \code{\link{is_installed}}, \code{\link{log_get}},
  \code{\link{pkgdetails_get}},
  \code{\link{string_replace}},
  \code{\link{templates_get}}, \code{\link{test}},
  \code{\link{to_basename}}
}
\concept{private}
