% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{install_test}
\alias{install_test}
\title{Test whether module can be installed}
\usage{
install_test(repo, tag)
}
\arguments{
\item{repo}{Module repo}

\item{tag}{Docker tag}
}
\value{
logical
}
\description{
Return TRUE if the outsider module successfully installs.
}
\seealso{
Other private: \code{\link{examples_test}},
  \code{\link{fnames_get}}, \code{\link{import_test}}
}
\concept{private}
