% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{is_docker_installed}
\alias{is_docker_installed}
\title{Check if Docker is installed}
\usage{
is_docker_installed()
}
\value{
Logical
}
\description{
Docker is required to run \code{outsider}. This function tests
whether Docker is installed.
}
\seealso{
Other private-check: \code{\link{is_docker_available}},
  \code{\link{is_docker_running}}
}
\concept{private-check}
