% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{is_docker_running}
\alias{is_docker_running}
\title{Check if Docker is running}
\usage{
is_docker_running()
}
\value{
Logical
}
\description{
Docker is required to run \code{outsider}. This function tests
whether Docker is running.
}
\seealso{
Other private-check: \code{\link{is_docker_available}},
  \code{\link{is_docker_installed}}
}
\concept{private-check}
