% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{is_filepath}
\alias{is_filepath}
\title{Is a filepath?}
\usage{
is_filepath(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Logical
}
\description{
Return TRUE or FALSE for whether character(s) is a valid
filepath.
}
\seealso{
Other private: \code{\link{default_log_set}},
  \code{\link{examples_test}}, \code{\link{file_create}},
  \code{\link{fnames_get}}, \code{\link{import_test}},
  \code{\link{install_test}}, \code{\link{install}},
  \code{\link{is_installed}}, \code{\link{log_get}},
  \code{\link{pkgdetails_get}},
  \code{\link{string_replace}},
  \code{\link{templates_get}}, \code{\link{test}},
  \code{\link{to_basename}}
}
\concept{private}
