% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{is_installed}
\alias{is_installed}
\title{Is module installed?}
\usage{
is_installed(repo)
}
\value{
logical
}
\description{
Return TRUE if module is installed.
}
\seealso{
Other private: \code{\link{default_log_set}},
  \code{\link{examples_test}}, \code{\link{file_create}},
  \code{\link{fnames_get}}, \code{\link{import_test}},
  \code{\link{install_test}}, \code{\link{install}},
  \code{\link{is_filepath}}, \code{\link{log_get}},
  \code{\link{pkgdetails_get}},
  \code{\link{string_replace}},
  \code{\link{templates_get}}, \code{\link{test}},
  \code{\link{to_basename}}
}
\concept{private}
