% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outsider.R
\name{is_outsider_ready}
\alias{is_outsider_ready}
\title{Is outsider ready to run?}
\usage{
is_outsider_ready()
}
\value{
logical
}
\description{
Return TRUE if modules can be installed and run. Provides
helpful messages on what may be missing.
}
