% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{is_repo_name}
\alias{is_repo_name}
\title{Is GitHub repo name valid?}
\usage{
is_repo_name(x)
}
\arguments{
\item{x}{Target name}
}
\description{
Calls error if repo name is invalid.
}
\seealso{
Other private-ids: \code{\link{ids_get}},
  \code{\link{pkgnm_to_img}}, \code{\link{pkgnm_to_prgm}},
  \code{\link{pkgnm_to_repo}}, \code{\link{repo_to_img}},
  \code{\link{repo_to_pkgnm}}
}
\concept{private-ids}
