% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launcher_class}
\alias{launcher_class}
\title{Construct launcher object}
\usage{
launcher_class(repo, cmd = NA, arglist = NULL, wd = NULL,
  files_to_send = NULL)
}
\arguments{
\item{repo}{Repository of the outsider module}

\item{cmd}{Command to be called in the container}

\item{arglist}{Arguments for command, character vector}

\item{wd}{Directory to which program generated files will be returned}

\item{files_to_send}{Files to be sent to container}
}
\value{
Character vector
}
\description{
Returns a launcher object. The launcher object describes a
outsider module's program and arguments. The object is generated every
time an outsider module program is called. It details the arguments of a
call, the command as well as the files to send to the docker container.
}
\details{
The outsider module runs a docker container that acts like a 
separate machine on the host computer. All the files necessary for the 
program to be run must be sent to the remote machine before the program
is called.
}
