% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_get}
\alias{log_get}
\title{Return log stream option}
\usage{
log_get(log = c("program_out", "program_err", "docker_out",
  "docker_err"))
}
\arguments{
\item{log}{Log stream}
}
\description{
Return the log stream setting for a given stream. If the stream
is not set, the function will return TRUE (i.e. prints to console).
}
\seealso{
Other private: \code{\link{build_status}},
  \code{\link{default_log_set}}, \code{\link{fnames_get}},
  \code{\link{is_docker_available}},
  \code{\link{is_filepath}}, \code{\link{om_search}},
  \code{\link{om_yaml}}, \code{\link{test_examples}},
  \code{\link{test_import}}, \code{\link{test_install}}
}
\concept{private}
