% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_set}
\alias{log_set}
\title{Set log streams for console output}
\usage{
log_set(log, val)
}
\arguments{
\item{log}{Output stream one of}

\item{val}{Either logcal, file or connection.}

\item{std_err}{if and where to direct child process STDERR.
See \code{\link[sys]{exec}}.}
}
\description{
Set if and where to send the console streams of the outsider
modules.
}
\details{
See \code{\link[sys]{exec}}.
}
