% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{module_check}
\alias{module_check}
\title{Check names and structure of a module}
\usage{
module_check(flpth = NULL)
}
\arguments{
\item{flpth}{File path to location of module}
}
\value{
Logical
}
\description{
Returns TRUE if all the names and structure of an outsider
module are correct.
}
\seealso{
Other developer: \code{\link{.arglist_get}},
  \code{\link{.arglist_parse}},
  \code{\link{.filestosend_get}},
  \code{\link{.travis_yaml}}, \code{\link{.wd_get}},
  \code{\link{module_identities}},
  \code{\link{module_skeleton}}, \code{\link{module_test}},
  \code{\link{outsider-class}}
}
\concept{developer}
