% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module_exists}
\alias{module_exists}
\title{Does module exist?}
\usage{
module_exists(repo)
}
\arguments{
\item{repo}{Module repo(s)}
}
\value{
Logical
}
\description{
Does the module(s) exist as a valid outsider module? Repo
must be a valid GitHub repository with an om.yaml and a passing
build status.
}
\seealso{
Other user: \code{\link{available}}, \code{\link{log_set}},
  \code{\link{module_help}}, \code{\link{module_import}},
  \code{\link{module_installed}},
  \code{\link{module_install}}, \code{\link{module_test}},
  \code{\link{module_uninstall}}
}
\concept{user}
