% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{module_functions}
\alias{module_functions}
\title{List the functions associated with a module}
\usage{
module_functions(repo)
}
\arguments{
\item{repo}{Module repo}
}
\value{
character
}
\description{
Return a vector of functions that can be imported from the
module.
}
\examples{
# NOT RUN
# library(outsider)
# # simplest repo
# repo <- 'dombennett/om..hello.world'
# # install
# module_install(repo = repo)
# hello_world <- module_import(fname = 'hello_world', repo = repo)
# # get help
# module_help(repo = repo)
# # import function
# hello_world <- module_import(fname = 'hello_world', repo = repo)
# # run function
# hello_world()
# 
# # uninstall
# module_uninstall(repo)
}
\seealso{
Other public: \code{\link{is_module_installed}},
  \code{\link{module_details}}, \code{\link{module_help}},
  \code{\link{module_import}},
  \code{\link{module_installed}},
  \code{\link{module_search}},
  \code{\link{module_uninstall}}
}
\concept{public}
