% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module_help}
\alias{module_help}
\title{Get help for outsider modules}
\usage{
module_help(repo, fname = NULL)
}
\arguments{
\item{repo}{Module repo}

\item{fname}{Function name}
}
\description{
Look up help files for specific outsider module functions or
whole modules.
}
