% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module_import}
\alias{module_import}
\title{Import functions from a module}
\usage{
module_import(fname, repo)
}
\arguments{
\item{fname}{Function name to import}

\item{repo}{Module repo}
}
\value{
Function
}
\description{
Import specific functions from an outsider module to the
Global Environment.
}
\details{
If program is successfully removed from your system, TRUE is
returned else FALSE.
}
