% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{module_skeleton}
\alias{module_skeleton}
\title{Generate a skeleton for a module}
\usage{
module_skeleton(flpth = getwd())
}
\arguments{
\item{flpth}{File path to location of where module will be created, default
current working directory.}
}
\value{
Logical
}
\description{
User is promted to supply key information, the function then
creates all the base files and folders to kickstart the development of a new
outsider module.
}
\details{
For more detailed information, see online.
}
\seealso{
Other developer: \code{\link{.arglist_get}},
  \code{\link{.arglist_parse}},
  \code{\link{.filestosend_get}},
  \code{\link{.travis_yaml}}, \code{\link{.wd_get}},
  \code{\link{module_check}},
  \code{\link{module_identities}},
  \code{\link{module_test}}, \code{\link{outsider-class}}
}
\concept{developer}
