% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{module_tags}
\alias{module_tags}
\title{Module tags}
\usage{
module_tags(repos)
}
\arguments{
\item{repos}{Character vector of outsider module repositories.}
}
\value{
tbl_df
}
\description{
Return tbl_df of module tags for a list of outsider
modules.
}
\seealso{
Other private-search: \code{\link{module_yaml}}
}
\concept{private-search}
