% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module_test}
\alias{module_test}
\title{Test an outsider module}
\usage{
module_test(repo)
}
\arguments{
\item{repo}{Module repo}
}
\value{
Logical
}
\description{
Ensure an outsider module builds correctly and meets certain
criteria.
}
