% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{module_versions}
\alias{module_versions}
\title{Module versions}
\usage{
module_versions(repos)
}
\arguments{
\item{repos}{Character vector of outsider module repositories.}
}
\value{
data.frame
}
\description{
Return data.frame of module versions for a list of outsider
modules.
}
\seealso{
Other private-search: \code{\link{module_yaml}}
}
\concept{private-search}
