% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{module_yaml}
\alias{module_yaml}
\title{Module YAML information}
\usage{
module_yaml(repos)
}
\arguments{
\item{repos}{Character vector of outsider module repositories.}
}
\value{
data.frame
}
\description{
Return data.frame of all YAML information of given outsider
module repos.
}
\seealso{
Other private: \code{\link{build_status}},
  \code{\link{default_log_set}},
  \code{\link{examples_test}}, \code{\link{fnames_get}},
  \code{\link{import_test}}, \code{\link{install_test}},
  \code{\link{is_docker_available}},
  \code{\link{is_filepath}}, \code{\link{log_get}},
  \code{\link{module_versions}}
}
\concept{private}
