% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{om_search}
\alias{om_search}
\title{Search for outsider modules}
\usage{
om_search()
}
\value{
list
}
\description{
Return list of all available outsider modules
}
\seealso{
Other private: \code{\link{build_status}},
  \code{\link{default_log_set}}, \code{\link{fnames_get}},
  \code{\link{is_docker_available}},
  \code{\link{is_filepath}}, \code{\link{log_get}},
  \code{\link{om_yaml}}, \code{\link{test_examples}},
  \code{\link{test_import}}, \code{\link{test_install}}
}
\concept{private}
