% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{om_yaml}
\alias{om_yaml}
\title{Module YAML information}
\usage{
om_yaml(repos)
}
\arguments{
\item{repos}{Character vector of outsider module repositories.}
}
\value{
data.frame
}
\description{
Return data.frame of all YAML information of given outsider
module repos.
}
\seealso{
Other private: \code{\link{build_status}},
  \code{\link{default_log_set}}, \code{\link{fnames_get}},
  \code{\link{is_docker_available}},
  \code{\link{is_filepath}}, \code{\link{log_get}},
  \code{\link{om_search}}, \code{\link{test_examples}},
  \code{\link{test_import}}, \code{\link{test_install}}
}
\concept{private}
