% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outsider.R
\docType{package}
\name{outsider}
\alias{outsider}
\alias{outsider-package}
\title{outsider: Install and run programs, outside of R, inside of R}
\description{
The outsider package facilitates the installation and running of external
software by interfacing with docker (\url{https://www.docker.com/}).
External software are contained within mini-R-packages, called "outsider
modules" and can be installed directly to a user's computer through GitHub
(\url{https://github.com/}). The outsider package comes with a series of
functions for identifying and installing these outsider modules ("user"
functions). The package also comes with helper functions for
developing new outsider modules for hosting additional external programs
and software ("developer" functions).
}
\details{
For more information visit the outsider website
(\url{https://antonellilab.github.io/outsider/}).
}
