% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{pkgdetails_get}
\alias{pkgdetails_get}
\title{Read the package description}
\usage{
pkgdetails_get(flpth)
}
\arguments{
\item{flpth}{Path to package}
}
\value{
logical
}
\description{
Return a list of all package details based on a package's
DESCRIPTION file.
}
\seealso{
Other private: \code{\link{default_log_set}},
  \code{\link{examples_test}}, \code{\link{file_create}},
  \code{\link{fnames_get}}, \code{\link{import_test}},
  \code{\link{install_test}}, \code{\link{install}},
  \code{\link{is_filepath}}, \code{\link{is_installed}},
  \code{\link{log_get}}, \code{\link{string_replace}},
  \code{\link{templates_get}}, \code{\link{test}},
  \code{\link{to_basename}}
}
\concept{private}
