% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{pkgnm_guess}
\alias{pkgnm_guess}
\title{Guess package name}
\usage{
pkgnm_guess(repo, call_error = TRUE)
}
\arguments{
\item{repo}{Repository (e.g. "username/repo") or package name associated
with module}

\item{call_error}{Call error if no package found? Default, TRUE.}
}
\value{
character(1)
}
\description{
Return package name from a repo name.
}
