% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{pkgnm_to_repo}
\alias{pkgnm_to_repo}
\title{Convert pkgnm to repo}
\usage{
pkgnm_to_repo(pkgnm)
}
\arguments{
\item{pkgnm}{Package name}
}
\value{
Logical
}
\description{
Converts the R package name of a module into its repo name.
}
