% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{read_yaml}
\alias{read_yaml}
\title{Safely read om.yaml}
\usage{
read_yaml(repo)
}
\arguments{
\item{repo}{GitHub repo}
}
\value{
list
}
\description{
Return list of 'program' and 'details'.
}
\seealso{
Other private-search: \code{\link{all_search}},
  \code{\link{build_status}}, \code{\link{repo_search}},
  \code{\link{tags}}, \code{\link{yaml}}
}
\concept{private-search}
