% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{repo_search}
\alias{repo_search}
\title{Search for repository}
\usage{
repo_search(repo)
}
\arguments{
\item{repo}{GitHub repo}
}
\value{
data.frame
}
\description{
Return GitHub API item for specific repository.
}
\seealso{
Other private-search: \code{\link{all_search}},
  \code{\link{build_status}}, \code{\link{read_yaml}},
  \code{\link{tags}}, \code{\link{yaml}}
}
\concept{private-search}
