% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{repo_to_img}
\alias{repo_to_img}
\title{Convert repo to image ID}
\usage{
repo_to_img(repo)
}
\arguments{
\item{repo}{Repo}
}
\value{
Logical
}
\description{
Drops .. in repo name to meet docker name requirements.
}
\seealso{
Other private-ids: \code{\link{ids_get}},
  \code{\link{pkgnm_to_prgm}}, \code{\link{pkgnm_to_repo}},
  \code{\link{repo_to_pkgnm}}
}
\concept{private-ids}
