% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{string_replace}
\alias{string_replace}
\title{Replace patterns in a string}
\usage{
string_replace(string, patterns, values)
}
\arguments{
\item{string}{Text}

\item{patterns}{Patterns to replace with values}

\item{values}{Values to be put in place}
}
\value{
character
}
\description{
For a given character string, replace patterns with values.
}
\seealso{
Other private: \code{\link{default_log_set}},
  \code{\link{examples_test}}, \code{\link{file_create}},
  \code{\link{fnames_get}}, \code{\link{import_test}},
  \code{\link{install_test}}, \code{\link{install}},
  \code{\link{is_filepath}}, \code{\link{is_installed}},
  \code{\link{log_get}}, \code{\link{pkgdetails_get}},
  \code{\link{templates_get}}, \code{\link{test}},
  \code{\link{to_basename}}
}
\concept{private}
