% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{tags}
\alias{tags}
\title{Module tags}
\usage{
tags(repos)
}
\arguments{
\item{repos}{Character vector of outsider module repositories.}
}
\value{
tbl_df
}
\description{
Return tbl_df of module tags for a list of outsider
modules.
}
\seealso{
Other private-search: \code{\link{all_search}},
  \code{\link{build_status}}, \code{\link{read_yaml}},
  \code{\link{repo_search}}, \code{\link{yaml}}
}
\concept{private-search}
