% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{templates_get}
\alias{templates_get}
\title{Retreive template files}
\usage{
templates_get()
}
\value{
character vector
}
\description{
Return template files for an outsider module.
}
\seealso{
Other private: \code{\link{default_log_set}},
  \code{\link{examples_test}}, \code{\link{file_create}},
  \code{\link{fnames_get}}, \code{\link{import_test}},
  \code{\link{install_test}}, \code{\link{install}},
  \code{\link{is_filepath}}, \code{\link{is_installed}},
  \code{\link{log_get}}, \code{\link{pkgdetails_get}},
  \code{\link{string_replace}}, \code{\link{test}},
  \code{\link{to_basename}}
}
\concept{private}
