% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{test}
\alias{test}
\title{Test a module}
\usage{
test(repo)
}
\arguments{
\item{repo}{Repository}
}
\value{
logical
}
\description{
Test an outsider module by making sure it installs,
imports and its examples run correctly.
}
\seealso{
Other private: \code{\link{default_log_set}},
  \code{\link{examples_test}}, \code{\link{file_create}},
  \code{\link{fnames_get}}, \code{\link{import_test}},
  \code{\link{install_test}}, \code{\link{install}},
  \code{\link{is_filepath}}, \code{\link{is_installed}},
  \code{\link{log_get}}, \code{\link{pkgdetails_get}},
  \code{\link{string_replace}},
  \code{\link{templates_get}}, \code{\link{to_basename}}
}
\concept{private}
