% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_examples}
\alias{test_examples}
\title{Run each example of an outsider module}
\usage{
test_examples(repo)
}
\arguments{
\item{repo}{Module repo}
}
\value{
logical
}
\description{
Return TRUE if all of the outsider module functions successfully
run.
}
\seealso{
Other private: \code{\link{build_status}},
  \code{\link{default_log_set}}, \code{\link{fnames_get}},
  \code{\link{is_docker_available}},
  \code{\link{is_filepath}}, \code{\link{log_get}},
  \code{\link{om_search}}, \code{\link{om_yaml}},
  \code{\link{test_import}}, \code{\link{test_install}}
}
\concept{private}
