% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_install}
\alias{test_install}
\title{Test whether module can be installed}
\usage{
test_install(repo)
}
\arguments{
\item{repo}{Module repo}
}
\value{
logical
}
\description{
Return TRUE if the outsider module successfully installs.
}
\seealso{
Other private: \code{\link{build_status}},
  \code{\link{default_log_set}}, \code{\link{fnames_get}},
  \code{\link{is_docker_available}},
  \code{\link{is_filepath}}, \code{\link{log_get}},
  \code{\link{om_search}}, \code{\link{om_yaml}},
  \code{\link{test_examples}}, \code{\link{test_import}}
}
\concept{private}
