% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{url_make}
\alias{url_make}
\title{Make download URL}
\usage{
url_make(username, repo, ref = "master", service = c("github",
  "bitbucket", "gitlab"))
}
\arguments{
\item{username}{Username of repo}

\item{repo}{Repository name}

\item{ref}{Reference (e.g. branch, tag)}

\item{service}{One of 'github', 'bitbucket' or 'gitlab'}
}
\value{
character(1)
}
\description{
Return a url of the likely download .tar.gz link for a
repository
}
