% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{yaml}
\alias{yaml}
\title{Module YAML information}
\usage{
yaml(repos)
}
\arguments{
\item{repos}{Character vector of outsider module repositories.}
}
\value{
tbl_df
}
\description{
Return tbl_df of all YAML information of given outsider
module repos.
}
\seealso{
Other private-search: \code{\link{all_search}},
  \code{\link{build_status}}, \code{\link{read_yaml}},
  \code{\link{repo_search}}, \code{\link{tags}}
}
\concept{private-search}
