% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{yaml_read}
\alias{yaml_read}
\title{Module YAML information}
\usage{
yaml_read(repos, service = c("github", "gitlab", "bitbucket"))
}
\arguments{
\item{repos}{Character vector of outsider module repositories on GitHub,
GitLab or BitBucket.}

\item{service}{Code-sharing service. Character.}
}
\value{
tbl_df
}
\description{
Return tbl_df of all YAML information of given outsider
module repos.
}
