% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitbucket.R
\name{bitbucket_search}
\alias{bitbucket_search}
\title{Search for outsider modules in bitbucket}
\usage{
bitbucket_search(...)
}
\value{
data.frame
}
\description{
Returns bitbucket API item results for outsider module search.
}
