% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{build_status}
\alias{build_status}
\title{Look-up details on program}
\usage{
build_status(id)
}
\arguments{
\item{id}{Program ID}
}
\value{
Logical
}
\description{
Is build passing? Returns either TRUE or FALSE.
}
\seealso{
Other check: \code{\link{is_docker_available}}
}
\concept{check}
