% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{build_status}
\alias{build_status}
\title{Look-up details on program}
\usage{
build_status(repo)
}
\arguments{
\item{repo}{GitHub repo}
}
\value{
Logical
}
\description{
Is build passing? Returns either TRUE or FALSE.
}
\seealso{
Other private-check: \code{\link{is_docker_installed}},
  \code{\link{is_docker_running}}
}
\concept{private-check}
