% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{build_status}
\alias{build_status}
\title{Look-up details on program}
\usage{
build_status(repo)
}
\arguments{
\item{repo}{GitHub repo}
}
\value{
Logical
}
\description{
Is build passing? Returns either TRUE or FALSE.
}
\seealso{
Other private-search: \code{\link{all_search}},
  \code{\link{read_yaml}}, \code{\link{repo_search}},
  \code{\link{tags}}, \code{\link{yaml}}
}
\concept{private-search}
