% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{default_log_set}
\alias{default_log_set}
\title{Set default log streams}
\usage{
default_log_set()
}
\description{
By default all streams are printed to console with the exception
of \code{docker_out}.
}
\seealso{
Other private: \code{\link{build_status}},
  \code{\link{fnames_get}},
  \code{\link{is_docker_available}},
  \code{\link{is_filepath}}, \code{\link{log_get}},
  \code{\link{om_search}}, \code{\link{om_yaml}},
  \code{\link{test_examples}}, \code{\link{test_import}},
  \code{\link{test_install}}
}
\concept{private}
