% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{default_log_set}
\alias{default_log_set}
\title{Set default log streams}
\usage{
default_log_set()
}
\description{
By default all streams are printed to console with the exception
of \code{docker_out}.
}
\seealso{
Other private: \code{\link{examples_test}},
  \code{\link{file_create}}, \code{\link{fnames_get}},
  \code{\link{import_test}}, \code{\link{install_test}},
  \code{\link{install}}, \code{\link{is_filepath}},
  \code{\link{is_installed}}, \code{\link{log_get}},
  \code{\link{pkgdetails_get}},
  \code{\link{string_replace}},
  \code{\link{templates_get}}, \code{\link{test}},
  \code{\link{to_basename}}
}
\concept{private}
