% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_build}
\alias{docker_build}
\title{Build a docker image}
\usage{
docker_build(img, url_or_path, tag = "latest")
}
\arguments{
\item{img}{Image name}

\item{url_or_path}{Dockerfile URL}

\item{tag}{Docker tag, default 'latest'}
}
\value{
Logical
}
\description{
Runs run \code{build} command.
}
\seealso{
Other private-docker: \code{\link{container-class}},
  \code{\link{docker_cmd}}, \code{\link{docker_cp}},
  \code{\link{docker_img_ls}}, \code{\link{docker_img_rm}},
  \code{\link{docker_ps_count}}, \code{\link{docker_pull}}
}
\concept{private-docker}
