% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{.arglist_get}
\alias{.arglist_get}
\title{Generate vector of arguments}
\usage{
.arglist_get(...)
}
\arguments{
\item{...}{Any number of arguments}
}
\value{
Character vector
}
\description{
Convert all the arguments passed to this function, including
those contained in '...', into character vector.
}
\examples{
library(outsider)
# return a character vector of all arguments provided
.arglist_get('any', 'number', 'of', 'arguments...', 'number', 'or', 'anything',
             1L, TRUE)
}
\seealso{
Other developer: \code{\link{.arglist_parse}},
  \code{\link{.dirpath_get}},
  \code{\link{.filestosend_get}},
  \code{\link{.module_check}},
  \code{\link{.module_identities}},
  \code{\link{.module_skeleton}},
  \code{\link{.module_test}}, \code{\link{.module_travis}},
  \code{\link{.wd_get}}, \code{\link{outsider-class}}
}
\concept{developer}
