% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{.docker_cp}
\alias{.docker_cp}
\title{Copy files to and from container}
\usage{
.docker_cp(origin, dest)
}
\arguments{
\item{origin}{Origin filepath}

\item{dest}{Destination filepath}
}
\value{
Logical
}
\description{
Copy files to and from running Docker container
}
\details{
Container foldes are indicated with
\code{[container_id]:[filepath]}.
}
\seealso{
Other developer: \code{\link{.docker_build}},
  \code{\link{.docker_cmd}}, \code{\link{.docker_exec}},
  \code{\link{.docker_img_rm}},
  \code{\link{.docker_ps_count}},
  \code{\link{.docker_start}}, \code{\link{.docker_stop}},
  \code{\link{.ids_get}}, \code{\link{.pkgnm_to_prgm}},
  \code{\link{.pkgnm_to_repo}}, \code{\link{.repo_to_img}},
  \code{\link{.repo_to_pkgnm}}, \code{\link{.run_nocopy}},
  \code{\link{.run_noreturn}}, \code{\link{.run_nosend}},
  \code{\link{.run}}, \code{\link{.travisyml_gen}}
}
\concept{developer}
