% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{.docker_exec}
\alias{.docker_exec}
\title{Execute in a docker container}
\usage{
.docker_exec(cntnr_id, ...)
}
\arguments{
\item{cntnr_id}{Container ID}

\item{...}{Command and arguments for container.}
}
\value{
Logical
}
\description{
Pass commands to a running docker container with the \code{exec}
command.
}
\seealso{
Other developer: \code{\link{.args_parse}},
  \code{\link{.docker_build}}, \code{\link{.docker_cmd}},
  \code{\link{.docker_cp}}, \code{\link{.docker_img_rm}},
  \code{\link{.docker_ps_count}},
  \code{\link{.docker_start}}, \code{\link{.docker_stop}},
  \code{\link{.ids_get}}, \code{\link{.pkgnm_to_prgm}},
  \code{\link{.pkgnm_to_repo}}, \code{\link{.repo_to_img}},
  \code{\link{.repo_to_pkgnm}}, \code{\link{.run}},
  \code{\link{.which_args_are_filepaths}}
}
\concept{developer}
