% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{.docker_ps_count}
\alias{.docker_ps_count}
\title{Count docker processes}
\usage{
.docker_ps_count()
}
\value{
Integer
}
\description{
Count the number of running docker containers.
}
\details{
Use this to avoid creating multiple containers with the same ID.
}
\seealso{
Other developer: \code{\link{.docker_build}},
  \code{\link{.docker_cmd}}, \code{\link{.docker_cp}},
  \code{\link{.docker_exec}}, \code{\link{.docker_img_rm}},
  \code{\link{.docker_start}}, \code{\link{.docker_stop}},
  \code{\link{.ids_get}}, \code{\link{.pkgnm_to_prgm}},
  \code{\link{.pkgnm_to_repo}}, \code{\link{.repo_to_img}},
  \code{\link{.repo_to_pkgnm}}, \code{\link{.run_nocopy}},
  \code{\link{.run_noreturn}}, \code{\link{.run_nosend}},
  \code{\link{.run}}, \code{\link{.travisyml_gen}}
}
\concept{developer}
