% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{.docker_stop}
\alias{.docker_stop}
\title{Stop and delete a docker container}
\usage{
.docker_stop(cntnr_id)
}
\arguments{
\item{cntnr_id}{Container ID}
}
\value{
Logical
}
\description{
Runs run \code{stop} and \code{rm} commands on running docker
containers.
}
\seealso{
Other developer: \code{\link{.args_parse}},
  \code{\link{.copy_from_docker}},
  \code{\link{.copy_to_docker}},
  \code{\link{.docker_build}}, \code{\link{.docker_cmd}},
  \code{\link{.docker_cp}}, \code{\link{.docker_exec}},
  \code{\link{.docker_img_rm}},
  \code{\link{.docker_ps_count}},
  \code{\link{.docker_start}}, \code{\link{.ids_get}},
  \code{\link{.pkgnm_to_prgm}},
  \code{\link{.pkgnm_to_repo}}, \code{\link{.repo_to_img}},
  \code{\link{.repo_to_pkgnm}}, \code{\link{.run}},
  \code{\link{.travisyml_gen}},
  \code{\link{.which_args_are_filepaths}}
}
\concept{developer}
