% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{.filestosend_get}
\alias{.filestosend_get}
\title{Determine which arguments are filepaths}
\usage{
.filestosend_get(arglist, wd = NULL)
}
\arguments{
\item{arglist}{Character vector of arguments}

\item{wd}{Working directory in which to look for files}
}
\value{
Character vector
}
\description{
Return filepaths from arguments. These filepaths can then be
used to identify files/folders for sending to the Docker container.
}
