% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{.filestosend_get}
\alias{.filestosend_get}
\title{Determine which arguments are filepaths}
\usage{
.filestosend_get(arglist, wd = NULL)
}
\arguments{
\item{arglist}{Character vector of arguments}

\item{wd}{Working directory in which to look for files}
}
\value{
Character vector
}
\description{
Return filepaths from arguments. These filepaths can then be
used to identify files/folders for sending to the Docker container.
}
\seealso{
Other developer: \code{\link{.arglist_get}},
  \code{\link{.arglist_parse}},
  \code{\link{.module_check}},
  \code{\link{.module_identities}},
  \code{\link{.module_skeleton}},
  \code{\link{.module_travis}}, \code{\link{.wd_get}},
  \code{\link{module_test}}, \code{\link{outsider-class}}
}
\concept{developer}
