% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outsider.R
\name{.run}
\alias{.run}
\alias{.run.outsider}
\title{Run an outsider module}
\usage{
\method{.run}{outsider}(x)
}
\arguments{
\item{x}{outsider object}
}
\value{
Logical
}
\description{
Run an outsider module using the outsider object.
}
\seealso{
Other developer: \code{\link{.travisyml_gen}}
}
\concept{developer}
