% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{.run}
\alias{.run}
\title{Run a command from a docker image}
\usage{
.run(pkgnm, files_to_send, ...)
}
\arguments{
\item{pkgnm}{Package name}

\item{files_to_send}{Filepaths on host to send to module container}

\item{...}{Command and arguments}
}
\value{
Logical
}
\description{
Run a command with arguments within a docker image. The function
will start a container, run the command and after completion stop and remove
the container.
}
\seealso{
Other developer: \code{\link{.args_parse}},
  \code{\link{.docker_build}}, \code{\link{.docker_cmd}},
  \code{\link{.docker_cp}}, \code{\link{.docker_exec}},
  \code{\link{.docker_img_rm}},
  \code{\link{.docker_ps_count}},
  \code{\link{.docker_start}}, \code{\link{.docker_stop}},
  \code{\link{.ids_get}}, \code{\link{.pkgnm_to_prgm}},
  \code{\link{.pkgnm_to_repo}}, \code{\link{.repo_to_img}},
  \code{\link{.repo_to_pkgnm}},
  \code{\link{.which_args_are_filepaths}}
}
\concept{developer}
