% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{.travisyml_gen}
\alias{.travisyml_gen}
\title{Generate travis file}
\usage{
.travisyml_gen(repo, dir = getwd())
}
\arguments{
\item{repo}{Repository}

\item{dir}{Directory in which to create .travis.yml}
}
\value{
Logical
}
\description{
Write .travis.yml to working directory.
}
\details{
All validated outsider modules must have a .travis.yml in their
repository. These .travis.yml must be generated using this function.
}
\seealso{
Other developer: \code{\link{.run}}
}
\concept{developer}
