% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{.wd_get}
\alias{.wd_get}
\title{Return working directory}
\usage{
.wd_get(arglist, key = NULL, i = NULL)
}
\arguments{
\item{arglist}{Arguments as character vector}

\item{key}{Argument key identifying the working directory, e.g. -wd}

\item{i}{Index of the working directory in the arguments, e.g. 1.}
}
\value{
Character
}
\description{
Utility function for determines the working directory from 
arglist. The workign directory can be determined from the arglist either by
a key:value or an index. For example, the working directory may be determined
by the key \code{-wd} in which case this function will identify whether this
key exists in the arglist and will return its corresponding value.
If no key or i is provided or found in the arguments, returns the
R session's working directory.
}
\seealso{
Other developer: \code{\link{.arglist_get}},
  \code{\link{.arglist_parse}},
  \code{\link{.filestosend_get}},
  \code{\link{.travis_yaml}}, \code{\link{module_check}},
  \code{\link{module_identities}},
  \code{\link{module_skeleton}}, \code{\link{module_test}},
  \code{\link{outsider-class}}
}
\concept{developer}
