% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{examples_test}
\alias{examples_test}
\title{Run each example of an outsider module}
\usage{
examples_test(repo)
}
\arguments{
\item{repo}{Module repo}
}
\value{
logical
}
\description{
Return TRUE if all of the outsider module functions successfully
run.
}
\seealso{
Other private: \code{\link{default_log_set}},
  \code{\link{file_create}}, \code{\link{fnames_get}},
  \code{\link{import_test}}, \code{\link{install_test}},
  \code{\link{install}}, \code{\link{is_filepath}},
  \code{\link{is_installed}}, \code{\link{log_get}},
  \code{\link{pkgdetails_get}},
  \code{\link{string_replace}},
  \code{\link{templates_get}}, \code{\link{test}},
  \code{\link{to_basename}}
}
\concept{private}
