% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{fnames_get}
\alias{fnames_get}
\title{Function names for module}
\usage{
fnames_get(repo)
}
\arguments{
\item{repo}{Module repo}
}
\value{
character vector
}
\description{
Return function names of all available functions for an
installed outsider modules
}
\seealso{
Other private: \code{\link{default_log_set}},
  \code{\link{examples_test}}, \code{\link{file_create}},
  \code{\link{import_test}}, \code{\link{install_test}},
  \code{\link{install}}, \code{\link{is_filepath}},
  \code{\link{is_installed}}, \code{\link{log_get}},
  \code{\link{pkgdetails_get}},
  \code{\link{string_replace}},
  \code{\link{templates_get}}, \code{\link{test}},
  \code{\link{to_basename}}
}
\concept{private}
