% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{install_test}
\alias{install_test}
\title{Test whether module can be installed}
\usage{
install_test(repo)
}
\arguments{
\item{repo}{Module repo}
}
\value{
logical
}
\description{
Return TRUE if the outsider module successfully installs.
}
\seealso{
Other private: \code{\link{build_status}},
  \code{\link{default_log_set}},
  \code{\link{examples_test}}, \code{\link{fnames_get}},
  \code{\link{import_test}},
  \code{\link{is_docker_available}},
  \code{\link{is_filepath}}, \code{\link{log_get}},
  \code{\link{module_versions}}, \code{\link{module_yaml}}
}
\concept{private}
