% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{is_docker_available}
\alias{is_docker_available}
\title{Check if Docker is installed and running}
\usage{
is_docker_available()
}
\description{
Raises an error if docker is not available.
}
\seealso{
Other private-check: \code{\link{is_docker_installed}},
  \code{\link{is_docker_running}}
}
\concept{private-check}
