% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_docker_available}
\alias{is_docker_available}
\title{Check if Docker is available}
\usage{
is_docker_available()
}
\value{
Logical
}
\description{
Docker is required to run \code{outsider}. This function tests
whether Docker is available.
}
\seealso{
Other check: \code{\link{build_status}}
}
\concept{check}
