% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{is_filepath}
\alias{is_filepath}
\title{Is a filepath?}
\usage{
is_filepath(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Logical
}
\description{
Return TRUE or FALSE for whether character(s) is a valid
filepath.
}
\seealso{
Other private: \code{\link{build_status}},
  \code{\link{default_log_set}}, \code{\link{fnames_get}},
  \code{\link{is_docker_available}}, \code{\link{log_get}},
  \code{\link{om_search}}, \code{\link{om_yaml}},
  \code{\link{test_examples}}, \code{\link{test_import}},
  \code{\link{test_install}}
}
\concept{private}
