% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{is_module_installed}
\alias{is_module_installed}
\title{Is module installed?}
\usage{
is_module_installed(repo)
}
\arguments{
\item{repo}{Module repo}
}
\value{
Logical
}
\description{
Uninstall outsider module and removes it from your docker
}
\details{
If program is successfully removed from your system, TRUE is
returned else FALSE.
}
\examples{
# NOT RUN
# library(outsider)
# # simplest repo
# repo <- 'dombennett/om..hello.world'
# # install
# module_install(repo = repo)
# hello_world <- module_import(fname = 'hello_world', repo = repo)
# # get help
# module_help(repo = repo)
# # import function
# hello_world <- module_import(fname = 'hello_world', repo = repo)
# # run function
# hello_world()
# 
# # uninstall
# module_uninstall(repo)
}
\seealso{
Other user: \code{\link{module_help}},
  \code{\link{module_import}},
  \code{\link{module_installed}},
  \code{\link{module_uninstall}}
}
\concept{user}
