% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{is_module_installed}
\alias{is_module_installed}
\title{Is module installed?}
\usage{
is_module_installed(repo)
}
\arguments{
\item{repo}{Module repo}
}
\value{
Logical
}
\description{
Check if a module is installed on your system.
}
\details{
Searches for \code{repo} among installed outsider modules. Returns
TRUE if found, else FALSE.
}
\examples{
library(outsider)
# NOT RUN (too slow for automated testing)
\dontrun{
  if (is_outsider_ready()) {
    # simplest repo
    repo <- 'dombennett/om..hello.world'
    # install
    module_install(repo = repo, force = TRUE, update = 'never')
    # is module_installed?
    (is_module_installed(repo = repo))
    # get help for package
    #module_help(repo = repo)
    # list functions available
    module_functions(repo = repo)
    # import
    hello_world <- module_import(fname = 'hello_world', repo = repo)
    # get help for function
    #module_help(repo = repo, fname = 'hello_world')
    # ?hello_world # also works
    # run function
    hello_world()
    
    # change verbosity settings
    # print nothing to console
    verbosity_set(show_program = FALSE, show_docker = FALSE)
    hello_world()
    # print everything to console
    verbosity_set(show_program = TRUE, show_docker = TRUE)
    hello_world()
    # write program output to a file
    verbosity_set(show_program = 'log.txt', show_docker = FALSE)
    hello_world()
    (readLines(con = 'log.txt'))
    file.remove('log.txt')
    
    # uninstall
    module_uninstall(repo)
  }
}
}
\seealso{
Other public: 
\code{\link{module_details}()},
\code{\link{module_functions}()},
\code{\link{module_help}()},
\code{\link{module_import}()},
\code{\link{module_installed}()},
\code{\link{module_search}()},
\code{\link{module_uninstall}()}
}
\concept{public}
