% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_set}
\alias{log_set}
\title{Set log streams for console output}
\usage{
log_set(log, val)
}
\arguments{
\item{log}{Output stream one of program_out, program_err, docker_out or
docker_err}

\item{val}{Either logcal, file or connection.}
}
\description{
Set if and where to send the console streams of the outsider
modules.
}
\details{
See \code{\link[sys]{exec}}.
}
\seealso{
Other user: \code{\link{module_details}},
  \code{\link{module_exists}}, \code{\link{module_help}},
  \code{\link{module_import}},
  \code{\link{module_installed}},
  \code{\link{module_install}},
  \code{\link{module_search}},
  \code{\link{module_uninstall}}
}
\concept{user}
