% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_set}
\alias{log_set}
\title{Set log streams for console output}
\usage{
log_set(log, val)
}
\arguments{
\item{log}{Output stream one of program_out, program_err, docker_out or
docker_err}

\item{val}{Either logical, file or connection.}
}
\description{
Set if and where to send the console streams of the outsider
modules.
}
\details{
See \code{\link[sys]{exec}}.
}
\examples{
library(outsider)
# install and import
repo <- 'dombennett/om..hello.world'
module_install(repo = repo)
hello_world <- module_import(fname = 'hello_world', repo = repo)

# control the log stream
# send output to file
tmpfl <- tempfile()
log_set(log = 'program_out', val = tmpfl)
hello_world()
(readLines(con = tmpfl))
file.remove(tmpfl)
# send docker and program output to console
log_set(log = 'program_out', val = TRUE)
log_set(log = 'docker_out', val = TRUE)
hello_world()

# uninstall
rm(hello_world)
module_uninstall(repo = repo)
}
\seealso{
Other user: \code{\link{module_details}},
  \code{\link{module_exists}}, \code{\link{module_help}},
  \code{\link{module_import}},
  \code{\link{module_installed}},
  \code{\link{module_install}},
  \code{\link{module_search}},
  \code{\link{module_uninstall}}
}
\concept{user}
