% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{module_details}
\alias{module_details}
\title{Look up details on module(s)}
\usage{
module_details(repo = NULL, service = c("github", "bitbucket",
  "gitlab"))
}
\arguments{
\item{repo}{Vector of one or more outsider module repositories, default NULL.}

\item{service}{Code-sharing service, e.g. GitHub}
}
\value{
tbl_df
}
\description{
Return a tbl_df of information for outsider module(s) for a
given code-sharing service. If \code{repo} is NULL, will return details on
all available modules.
}
\details{
Module details in tibble format include: repository name
(user/repo), last time repo was updated, number of watchers (or stars in the
case of GitLab), url to web presence, names of tagged versions.
}
\examples{
library(outsider)
# return table of ALL available modules on GitHub
# NOT RUN - takes too long
# (available_modules <- module_search())

# look-up whether specific module exists
# NOT RUN
# repo <- 'dombennett/om..goodbye.world'
# (suppressWarnings(module_exists(repo = repo)))
repo <- 'dombennett/om..hello.world'
#(module_exists(repo = repo))
# get details on specific module(s)
#(module_details(repo = repo))
}
