% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{module_details}
\alias{module_details}
\title{Look up details on module(s)}
\usage{
module_details(repo = NULL)
}
\arguments{
\item{repo}{Vector of one or more outsider module repositories, default NULL.}
}
\value{
tbl_df
}
\description{
Return a tbl_df of information for outsider module(s).
If \code{repo} is NULL, will return details on all available modules.
}
\examples{
library(outsider)
# return table of ALL available modules on GitHub
# NOT RUN - takes too long
# (available_modules <- module_search())

# look-up whether specific module exists
# NOT RUN
# repo <- 'dombennett/om..goodbye.world'
# (suppressWarnings(module_exists(repo = repo)))
repo <- 'dombennett/om..hello.world'
(module_exists(repo = repo))
# get details on specific module(s)
(module_details(repo = repo))
}
\seealso{
Other user: \code{\link{is_module_installed}},
  \code{\link{module_exists}}, \code{\link{module_help}},
  \code{\link{module_import}},
  \code{\link{module_installed}},
  \code{\link{module_install}},
  \code{\link{module_search}},
  \code{\link{module_uninstall}}
}
\concept{user}
