% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module_details}
\alias{module_details}
\title{Look up details on module(s)}
\usage{
module_details(repo)
}
\arguments{
\item{repo}{Vector of one or more outsider module repositories}
}
\value{
data.frame
}
\description{
Return a data.frame of information for outsider module(s).
}
\seealso{
Other user: \code{\link{log_set}},
  \code{\link{module_exists}}, \code{\link{module_help}},
  \code{\link{module_import}},
  \code{\link{module_installed}},
  \code{\link{module_install}},
  \code{\link{module_search}}, \code{\link{module_test}},
  \code{\link{module_uninstall}}
}
\concept{user}
