% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{module_exists}
\alias{module_exists}
\title{Does module exist?}
\usage{
module_exists(repo)
}
\arguments{
\item{repo}{Module repo(s)}
}
\value{
Logical
}
\description{
Does the module(s) exist as a valid outsider module? Repo
must be a valid GitHub repository with an om.yaml and a passing
build status.
}
\examples{
library(outsider)
# return table of ALL available modules on GitHub
# NOT RUN - takes too long
# (available_modules <- module_search())

# look-up whether specific module exists
# NOT RUN
# repo <- 'dombennett/om..goodbye.world'
# (suppressWarnings(module_exists(repo = repo)))
repo <- 'dombennett/om..hello.world'
(module_exists(repo = repo))
# get details on specific module(s)
(module_details(repo = repo))
}
\seealso{
Other user: \code{\link{is_module_installed}},
  \code{\link{module_details}}, \code{\link{module_help}},
  \code{\link{module_import}},
  \code{\link{module_installed}},
  \code{\link{module_install}},
  \code{\link{module_search}},
  \code{\link{module_uninstall}}
}
\concept{user}
