% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{module_help}
\alias{module_help}
\title{Get help for outsider modules}
\usage{
module_help(repo, fname = NULL)
}
\arguments{
\item{repo}{Module repo}

\item{fname}{Function name}
}
\description{
Look up help files for specific outsider module functions or
whole modules.
}
\examples{
# NOT RUN
# library(outsider)
# # simplest repo
# repo <- 'dombennett/om..hello.world'
# # install
# module_install(repo = repo)
# hello_world <- module_import(fname = 'hello_world', repo = repo)
# # get help
# module_help(repo = repo)
# # import function
# hello_world <- module_import(fname = 'hello_world', repo = repo)
# # run function
# hello_world()
# 
# # uninstall
# module_uninstall(repo)
}
\seealso{
Other user: \code{\link{is_module_installed}},
  \code{\link{module_functions}},
  \code{\link{module_import}},
  \code{\link{module_installed}},
  \code{\link{module_uninstall}}
}
\concept{user}
