% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{module_import}
\alias{module_import}
\title{Import functions from a module}
\usage{
module_import(fname, repo)
}
\arguments{
\item{fname}{Function name to import}

\item{repo}{Module repo}
}
\value{
Function
}
\description{
Import specific functions from an outsider module to the
Global Environment.
}
\examples{
library(outsider)
if (is_outsider_ready()) {
  # simplest repo
  repo <- 'dombennett/om..hello.world'

  # is module_installed?
  if (is_module_installed(repo = repo)) {

    # get help for package
    module_help(repo = repo)
    
    # list functions available
    module_functions(repo = repo)
    
    # import
    hello_world <- module_import(fname = 'hello_world', repo = repo)
    
    # get help for function
    module_help(repo = repo, fname = 'hello_world')
    # also works
    ?hello_world

    # run function
    hello_world()
    
    # change verbosity settings
    
    # print nothing to console
    verbosity_set(show_program = FALSE, show_docker = FALSE)
    hello_world()
    
    # print everything to console
    verbosity_set(show_program = TRUE, show_docker = TRUE)
    hello_world()
    
    # write program output to a file
    log_file <- tempfile()
    verbosity_set(show_program = log_file, show_docker = FALSE)
    
    hello_world()
    
    (readLines(con = log_file))
    
    # Clean up
    file.remove(log_file)
  }
}
}
\seealso{
Other public: 
\code{\link{is_module_installed}()},
\code{\link{module_details}()},
\code{\link{module_functions}()},
\code{\link{module_help}()},
\code{\link{module_installed}()},
\code{\link{module_install}()},
\code{\link{module_search}()},
\code{\link{module_uninstall}()}
}
\concept{public}
