% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{module_import}
\alias{module_import}
\title{Import functions from a module}
\usage{
module_import(fname, repo)
}
\arguments{
\item{fname}{Function name to import}

\item{repo}{Module repo}
}
\value{
Function
}
\description{
Import specific functions from an outsider module to the
Global Environment.
}
\details{
If program is successfully removed from your system, TRUE is
returned else FALSE.
}
\examples{
# NOT RUN
# library(outsider)
# # simplest repo
# repo <- 'dombennett/om..hello.world'
# # install
# module_install(repo = repo)
# hello_world <- module_import(fname = 'hello_world', repo = repo)
# # get help
# module_help(repo = repo)
# # import function
# hello_world <- module_import(fname = 'hello_world', repo = repo)
# # run function
# hello_world()
# 
# # uninstall
# module_uninstall(repo)
}
\seealso{
Other user: \code{\link{is_module_installed}},
  \code{\link{module_help}},
  \code{\link{module_installed}},
  \code{\link{module_uninstall}}
}
\concept{user}
