% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{module_install}
\alias{module_install}
\title{Install an outsider module}
\usage{
module_install(repo, tag = "latest")
}
\arguments{
\item{repo}{Module repo}

\item{tag}{Module version, default latest}
}
\value{
Logical
}
\description{
Install a module
}
\seealso{
Other user: \code{\link{log_set}},
  \code{\link{module_details}},
  \code{\link{module_exists}}, \code{\link{module_help}},
  \code{\link{module_import}},
  \code{\link{module_installed}},
  \code{\link{module_search}},
  \code{\link{module_uninstall}}
}
\concept{user}
