% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{module_install}
\alias{module_install}
\title{Install an outsider module}
\usage{
module_install(repo = NULL, url = NULL, filepath = NULL,
  tag = "latest", manual = FALSE, verbose = FALSE, force = FALSE)
}
\arguments{
\item{repo}{Module repo, character.}

\item{url}{URL to downloadable .tar.gz of module, character.}

\item{filepath}{Filepath to uncompressed directory of module, character.}

\item{tag}{Module version, default latest. Character.}

\item{manual}{Build the docker image? Default FALSE. Logical.}

\item{verbose}{Be verbose? Default FALSE.}

\item{force}{Ignore warnings and install anyway? Default FALSE.}
}
\value{
Logical
}
\description{
Install a module
}
\examples{
# NOT RUN
# library(outsider)
# # simplest repo
# repo <- 'dombennett/om..hello.world'
# # install
# module_install(repo = repo)
# hello_world <- module_import(fname = 'hello_world', repo = repo)
# # get help
# module_help(repo = repo)
# # import function
# hello_world <- module_import(fname = 'hello_world', repo = repo)
# # run function
# hello_world()
# 
# # uninstall
# module_uninstall(repo)
}
