% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module_installed}
\alias{module_installed}
\title{Is module installed?}
\usage{
module_installed(repo)
}
\arguments{
\item{repo}{Module repo(s)}
}
\value{
Logical
}
\description{
Test whether an outsider module is installed.
}
\seealso{
Other user: \code{\link{log_set}},
  \code{\link{module_details}},
  \code{\link{module_exists}}, \code{\link{module_help}},
  \code{\link{module_import}},
  \code{\link{module_install}},
  \code{\link{module_search}}, \code{\link{module_test}},
  \code{\link{module_uninstall}}
}
\concept{user}
