% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{module_installed}
\alias{module_installed}
\title{Which outsider modules are installed?}
\usage{
module_installed()
}
\arguments{
\item{show_images}{Look-up the module images? Default FALSE.}
}
\value{
tbl_df
}
\description{
Returns tbl_df of details for all outsider modules
installed on the user's computer.
}
\examples{
library(outsider)
# check what modules are installed
(module_installed())
}
\seealso{
Other user: \code{\link{is_module_installed}},
  \code{\link{module_details}},
  \code{\link{module_exists}}, \code{\link{module_help}},
  \code{\link{module_import}},
  \code{\link{module_install}},
  \code{\link{module_search}},
  \code{\link{module_uninstall}}
}
\concept{user}
