% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{module_search}
\alias{module_search}
\title{Search for available outsider modules}
\usage{
module_search(service = c("github", "gitlab"))
}
\arguments{
\item{service}{Code-sharing service, e.g. GitHub}
}
\value{
Character vector
}
\description{
Return a list of available outsider modules. (Not possible for
BitBucket.)
}
\examples{
library(outsider)
# return table of ALL available modules on GitHub
# NOT RUN - takes too long
# (available_modules <- module_search())

# look-up specific modules
repo <- 'dombennett/om..goodbye.world'
(suppressWarnings(module_details(repo = repo))) # no module exists, expect warning
repo <- 'dombennett/om..hello.world'
(module_details(repo = repo))
}
\seealso{
Other public: \code{\link{is_module_installed}},
  \code{\link{module_details}},
  \code{\link{module_functions}},
  \code{\link{module_help}}, \code{\link{module_import}},
  \code{\link{module_installed}},
  \code{\link{module_uninstall}}
}
\concept{public}
