% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{module_search}
\alias{module_search}
\title{Search for available outsider modules}
\usage{
module_search(service = c("github", "bitbucket", "gitlab"))
}
\arguments{
\item{service}{Code-sharing service, e.g. GitHub}
}
\value{
Character vector
}
\description{
Return a list of available outsider modules.
}
\examples{
library(outsider)
# return table of ALL available modules on GitHub
# NOT RUN - takes too long
# (available_modules <- module_search())

# look-up whether specific module exists
# NOT RUN
# repo <- 'dombennett/om..goodbye.world'
# (suppressWarnings(module_exists(repo = repo)))
repo <- 'dombennett/om..hello.world'
#(module_exists(repo = repo))
# get details on specific module(s)
#(module_details(repo = repo))
}
