% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{module_search}
\alias{module_search}
\title{Search for available outsider modules}
\usage{
module_search()
}
\value{
Character vector
}
\description{
Return a list of available outsider modules.
}
\seealso{
Other user: \code{\link{log_set}},
  \code{\link{module_details}},
  \code{\link{module_exists}}, \code{\link{module_help}},
  \code{\link{module_import}},
  \code{\link{module_installed}},
  \code{\link{module_install}},
  \code{\link{module_uninstall}}
}
\concept{user}
