% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module_test}
\alias{module_test}
\title{Test an outsider module}
\usage{
module_test(repo, verbose = FALSE)
}
\arguments{
\item{repo}{Module repo}

\item{verbose}{Print docker and program info to console}
}
\value{
Logical
}
\description{
Ensure an outsider module builds, imports correctly and all
its functions successfully complete.
}
\details{
Success or fail, the module is uninstalled from the machine after
the test is run.
}
\seealso{
Other user: \code{\link{available}}, \code{\link{log_set}},
  \code{\link{module_exists}}, \code{\link{module_help}},
  \code{\link{module_import}},
  \code{\link{module_installed}},
  \code{\link{module_install}},
  \code{\link{module_uninstall}}
}
\concept{user}
