% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module_uninstall}
\alias{module_uninstall}
\title{Uninstall and remove a module}
\usage{
module_uninstall(repo)
}
\arguments{
\item{repo}{Module repo}
}
\value{
Logical
}
\description{
Uninstall outsider module and removes it from your docker
}
\details{
If program is successfully removed from your system, TRUE is
returned else FALSE.
}
