% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{module_yaml}
\alias{module_yaml}
\title{Module YAML information}
\usage{
module_yaml(repos)
}
\arguments{
\item{repos}{Character vector of outsider module repositories.}
}
\value{
tbl_df
}
\description{
Return tbl_df of all YAML information of given outsider
module repos.
}
\seealso{
Other private-search: \code{\link{module_tags}}
}
\concept{private-search}
