% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outsider.R
\docType{package}
\name{outsider}
\alias{outsider}
\title{outsider: Install and run programs, outside of R, inside of R}
\description{
The outsider package facilitates the installation and running of external
software by interfacing with docker (\url{https://www.docker.com/}).
External software are contained within mini-R-packages, called "outsider
modules" and can be installed directly to a user's computer through online
code-sharing services such as GitHub (\url{https://github.com/}). The
outsider package comes with a series of functions for identifying,
installing and importing these outsider modules.
}
\details{
For more information visit the outsider website
(\url{https://antonellilab.github.io/outsider/}).
}
