% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{pkgnm_to_repo}
\alias{pkgnm_to_repo}
\title{Convert pkgnm to repo}
\usage{
pkgnm_to_repo(pkgnm)
}
\arguments{
\item{pkgnm}{Package name}
}
\value{
Logical
}
\description{
Converts the R package name of a module into its repo name.
}
\seealso{
Other private-ids: \code{\link{ids_get}},
  \code{\link{is_repo_name}}, \code{\link{pkgnm_to_img}},
  \code{\link{pkgnm_to_prgm}}, \code{\link{repo_to_img}},
  \code{\link{repo_to_pkgnm}}
}
\concept{private-ids}
