% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{repo_to_img}
\alias{repo_to_img}
\title{Convert repo to image ID}
\usage{
repo_to_img(repo)
}
\arguments{
\item{repo}{Repo}
}
\value{
Logical
}
\description{
Drops .. in repo name to meet docker name requirements.
}
