% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identities.R
\name{repo_to_pkgnm}
\alias{repo_to_pkgnm}
\title{Convert repo to pkgnm}
\usage{
repo_to_pkgnm(repo)
}
\arguments{
\item{repo}{Repo}
}
\value{
Logical
}
\description{
Converts the repo name of a module into its R package name.
}
