% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outsider.R
\name{ssh_setup}
\alias{ssh_setup}
\title{Setup SSH}
\usage{
ssh_setup(session)
}
\arguments{
\item{session}{\code{ssh} session, see \code{\link[ssh]{ssh_connect}}}
}
\value{
logical
}
\description{
Send all outsider commands to an external host. Provide an
\code{ssh} session to this function and all subsequent commands will be run
on the host rather than the local machine. When finished it is always good
practice to disconnect from the remote host by running \code{ssh_teardown}.
}
\examples{
library(outsider)

# # NOT RUN
# session <- ssh::ssh_connect(host = "[IP ADDRESS]")
# ssh_setup(session)
# # (Run outsider commands)
# ssh_teardown()
}
