% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outsider.R
\name{ssh_teardown}
\alias{ssh_teardown}
\title{Teardown SSH}
\usage{
ssh_teardown()
}
\value{
logical
}
\description{
Disconnect from a remote host and stop commands being
transferred.
}
\examples{
library(outsider)

# # NOT RUN
# session <- ssh::ssh_connect(host = "[IP ADDRESS]")
# ssh_setup(session)
# # (Run outsider commands)
# ssh_teardown()
}
