% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis.R
\name{travis_build_status}
\alias{travis_build_status}
\title{Check Travis build status}
\usage{
travis_build_status(repo)
}
\arguments{
\item{repo}{repo}
}
\value{
Logical
}
\description{
Is build passing on travis? Returns either TRUE or FALSE.
}
\details{
For GitHub-based repositories only.
}
