% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{user_warn}
\alias{user_warn}
\title{Warn users on the dangers of outsider modules}
\usage{
user_warn(pkgnm)
}
\arguments{
\item{pkgnm}{Package name}
}
\value{
Logical
}
\description{
Warn users on the dangers of installing an outsider module
whose origin is potentially unknown.
}
\details{
Prints additional info to screen based on module YAML file.
}
